enable ruff in tests folder move all pytest fixtures into conftest.py refactor test code to pass ruff
