+++ tests/phantoms/test_coil_sensitivities.py
@@ -12,7 +12,6 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import torch
 
 from mrpro.data import SpatialDimension
 from mrpro.phantoms.coils import birdcage_2d
